First refactoring of config logic into separate class
