+++ lib/update_repo.rb
@@ -1,5 +1,6 @@
 require 'update_repo/version'
 require 'update_repo/helpers'
+require 'update_repo/cmd_config'
 require 'yaml'
 require 'colorize'
 require 'confoog'
@@ -24,18 +25,10 @@ module UpdateRepo
                    start_time: 0 }
       @summary = { processed: 'green', updated: 'cyan', skipped: 'yellow',
                    failed: 'red' }
-      # read the options from Trollop and store in temp variable.
-      # we do it this way around otherwise if configuration file is missing it
-      # gives the error messages even on '--help' and '--version'
-      temp_opt = set_options
-      # @config - Class. Reads the configuration from a file in YAML format and
-      # allows easy access to the configuration data
-      @config = Confoog::Settings.new(filename: CONFIG_FILE,
-                                      prefix: 'update_repo',
-                                      autoload: true, autosave: false)
-      # store the command line variables in a configuration variable
-      @config['cmd'] = temp_opt
-      config_error unless @config.status[:errors] == Status::INFO_FILE_LOADED
+      # create a new instance of the CmdConfig class then read the config var
+      @cmd = CmdConfig.new
+      # set up the logfile if needed
+      setup_logfile if logging?
     end
 
     # This function will perform the required actions to traverse the Repo.
@@ -43,12 +36,12 @@ module UpdateRepo
     #   walk_repo = UpdateRepo::WalkRepo.new
     #   walk_repo.start
     def start
-      String.disable_colorization = true unless param_set('color')
+      String.disable_colorization = true unless @cmd.param_set('color')
       # make sure we dont have bad cmd-line parameter combinations ...
-      check_params
+      @cmd.check_params
       # print out our header ...
       show_header unless dumping? || importing?
-      @config['location'].each do |loc|
+      config['location'].each do |loc|
         recurse_dir(loc)
       end
       # print out an informative footer ...
@@ -57,71 +50,19 @@ module UpdateRepo
 
     private
 
-    def check_params
-      setup_logfile if logging?
-      if dumping? && importing?
-        Trollop.die 'Sorry, you cannot specify both --dump and --import '.red
-      end
-      if importing?
-        Trollop.die "Sorry 'Import' functionality is not implemented yet".red
-      end
+    def config
+      @cmd.getconfig
     end
 
     def setup_logfile
-      filename = if param_set('timestamp')
+      filename = if @cmd.param_set('timestamp')
                    'updaterepo-' + Time.new.strftime('%y%m%d-%H%M%S') + '.log'
                  else
                    'updaterepo.log'
                  end
-      # filename = 'updaterepo' + prefix + '.log'
       @logfile = File.open(filename, 'w')
       @logfile.sync = true
     end
-
-    # Determine options from the command line and configuration file. Command
-    # line takes precedence
-    def param_set(option)
-      @config['cmd'][option.to_sym] || @config[option]
-    end
-
-    def config_error
-      if @config.status[:errors] == Status::ERR_CANT_LOAD
-        print_log 'Note that the the default configuration file was '.red,
-                  "changed to ~/#{CONFIG_FILE} from v0.4.0 onwards\n\n".red
-      end
-      exit 1
-    end
-
-    # rubocop:disable Metrics/MethodLength
-    # rubocop:disable Metrics/LineLength
-    def set_options
-      Trollop.options do
-        version "update_repo version #{VERSION} (C)2016 G. Ramsay\n"
-        banner <<-EOS
-
-Keep multiple local Git-Cloned Repositories up to date with one command.
-
-Usage:
-      update_repo [options]
-
-Options are not required. If none are specified then the program will read from
-the standard configuration file (~/#{CONFIG_FILE}) and automatically update the
-specified Repositories.
-
-Options:
-EOS
-        opt :color, 'Use colored output', default: true
-        opt :dump, 'Dump a list of Directories and Git URL\'s to STDOUT in CSV format', default: false
-        opt :prune, "Number of directory levels to remove from the --dump output.\nOnly valid when --dump or -d specified", default: 0
-        opt :import, "Import a previous dump of directories and Git repository URL's,\n(created using --dump) then proceed to clone them locally.", default: false
-        opt :log, "Create a logfile of all program output to './update_repo.log'. Any older logs will be overwritten.", default: false
-        opt :timestamp, 'Timestamp the logfile instead of overwriting. Does nothing unless the --log option is also specified.', default: false
-        # opt :quiet, 'Only minimal output to the terminal', default: false
-        # opt :silent, 'Completely silent, no output to terminal at all.', default: false
-      end
-    end
-    # rubocop:enable Metrics/MethodLength
-    # rubocop:enable  Metrics/LineLength
 
     # take each directory contained in the Repo directory, if it is detected as
     # a Git repository then update it (or as directed by command line)
@@ -143,7 +84,7 @@ EOS
     # @param dir [string] Directory to be checked
     # @return [boolean] True if this is NOT an exception, False otherwise
     def notexception?(dir)
-      !@config['exceptions'].include?(File.basename(dir))
+      !config['exceptions'].include?(File.basename(dir))
     end
 
     # Display a simple header to the console
@@ -155,8 +96,8 @@ EOS
       # unless we are dumping the repo information
       print_log "\nGit Repo update utility (v", VERSION, ')',
                 " \u00A9 Grant Ramsay <seapagan@gmail.com>\n"
-      print_log "Using Configuration from '#{@config.config_path}'\n"
-      # print_log "Command line is : #{@config['cmd']}\n"
+      print_log "Using Configuration from '#{config.config_path}'\n"
+      # print_log "Command line is : #{config['cmd']}\n"
       # list out the locations that will be searched
       list_locations
       # list any exceptions that we have from the config file
@@ -186,7 +127,7 @@ EOS
     end
 
     def list_exceptions
-      exceptions = @config['exceptions']
+      exceptions = config['exceptions']
       if exceptions
         print_log "\nExclusions:".underline, ' ',
                   exceptions.join(', ').yellow, "\n"
@@ -195,7 +136,7 @@ EOS
 
     def list_locations
       print_log "\nRepo location(s):\n".underline
-      @config['location'].each do |loc|
+      config['location'].each do |loc|
         print_log '-> ', loc.cyan, "\n"
       end
     end
@@ -244,8 +185,23 @@ EOS
     def dump_repo(dir)
       Dir.chdir(dir.chomp!('/')) do
         repo_url = `git config remote.origin.url`.chomp
-        print_log "#{trunc_dir(dir, @config['cmd'][:prune])},#{repo_url}\n"
+        print_log "#{trunc_dir(dir, config['cmd'][:prune])},#{repo_url}\n"
       end
+    end
+
+    # true if we are dumping the file structure and git urls.
+    def dumping?
+      @cmd.param_set('dump')
+    end
+
+    # true if we are importing a previously dumped list of Git repos.
+    def importing?
+      @cmd.param_set('import')
+    end
+
+    # true if we are logging to file.
+    def logging?
+      @cmd.param_set('log')
     end
   end
 end
