--- lib/update_repo/cmdline.rb
@@ -1,10 +1,0 @@
-require 'update_repo/version'
-require 'update_repo/helpers'
-
-# this class will encapsulate the command line fucntionality and processing
-module UpdateRepo
-
-  class CmdLine
-    include Helpers
-  end
-end
